Don't use uint8_t type.

+++ libhotp/hotp.c
@@ -232,7 +232,7 @@ hotp_generate_otp (const char *secret,
   }
 
   {
-    uint8_t offset = hs[sizeof (hs) - 1] & 0x0f;
+    size_t offset = hs[sizeof (hs) - 1] & 0x0f;
 
     S = (((hs[offset] & 0x7f) << 24)
 	 | ((hs[offset + 1] & 0xff) << 16)
